\name{alphabet-package}
\alias{alphabet-package}
\alias{alphabet}
\docType{package}
\title{
\packageTitle{alphabet}
}
\description{
\packageDescription{alphabet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{alphabet}
\packageIndices{alphabet}
call alphabet(i)
}
\author{
\packageAuthor{alphabet}

Maintainer: \packageMaintainer{alphabet}
}
\references{

\keyword{ package }
\seealso{
}
\examples{
~~ simple examples of the most important functions ~~
}
